Move set up of watch to just before sending emails
